+++ jctools-core/src/main/java/org/jctools/maps/NonBlockingIdentityHashMap.java
@@ -78,8 +78,8 @@ import static org.jctools.util.UnsafeAccess.fieldOffset;
  * @author Cliff Click
  * @param <TypeK> the type of keys maintained by this map
  * @param <TypeV> the type of mapped values
- * 
- *  @author Prashant Deva 
+ *
+ *  @author Prashant Deva
  *  Modified from original NonBlockingHashMap to use identity equality.
  *  Uses System.identityHashCode() to calculate hashMap.
  *  Key equality is compared using '=='.
@@ -97,7 +97,7 @@ public class NonBlockingIdentityHashMap<TypeK, TypeV>
   private static final int _Oscale = UNSAFE.arrayIndexScale(Object[].class);
   private static long rawIndex(final Object[] ary, final int idx) {
     assert idx >= 0 && idx < ary.length;
-    return _Obase + idx * _Oscale;
+    return _Obase + (idx * (long)_Oscale);
   }
 
   // --- Setup to use Unsafe
@@ -279,17 +279,17 @@ public class NonBlockingIdentityHashMap<TypeK, TypeV>
 
   /** Returns the number of key-value mappings in this map.
    *  @return the number of key-value mappings in this map */
-  @Override 
+  @Override
   public int     size       ( )                       { return chm(_kvs).size(); }
   /** Returns <tt>size() == 0</tt>.
    *  @return <tt>size() == 0</tt> */
-  @Override 
+  @Override
   public boolean isEmpty    ( )                       { return size() == 0;      }
 
   /** Tests if the key in the table using the <tt>equals</tt> method.
    * @return <tt>true</tt> if the key is in the table using the <tt>equals</tt> method
    * @throws NullPointerException if the specified key is null  */
-  @Override 
+  @Override
   public boolean containsKey( Object key )            { return get(key) != null; }
 
   /** Legacy method testing if some key maps into the specified value in this
@@ -825,7 +825,7 @@ public class NonBlockingIdentityHashMap<TypeK, TypeV>
         //synchronized( this ) { wait(8*megs); }         // Timeout - we always wakeup
         // For now, sleep a tad and see if the 2 guys already trying to make
         // the table actually get around to making it happen.
-        try { Thread.sleep(8*megs); } catch( Exception e ) { }
+        try { Thread.sleep(8l*megs); } catch( Exception e ) { }
       }
       // Last check, since the 'new' below is expensive and there is a chance
       // that another thread slipped in a new thread while we ran the heuristic.
